# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Node.js API model",
    "description": "Node.js Express API Microservice with Mongo",
    "contact": {
      "email": "david.a.m@live.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api"
    },
    {
      "url": "https://localhost:8080/api"
    }
  ],
...
  "paths": {
...
    "/uploads": {
      "post": {
        "tags": ["Uploads"],
        "summary": "Upload files to server",
        "description": "Upload a file to server for products or users.",
        "operationId": "uploadFiles",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Error uploading file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      }
    },
    "/uploads/{collection}/{id}": {
      "get": {
        "tags": ["Uploads"],
        "summary": "Get files from server or Cloudinary",
        "description": "Get files from server or Cloudinary for a defined collection.",
        "operationId": "getFiles",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "description": "The collection to search the term",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Element ID",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "image": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID or collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "image": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "501": {
            "description": "Collection not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Error getting file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "put": {
        "tags": ["Uploads"],
        "summary": "Update a file in server",
        "description": "Upload a file to server for products or users.",
        "operationId": "updateLocalImage",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "description": "The collection to search the term",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Element ID",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/Product" },
                    { "$ref": "#/components/schemas/Category" },
                    { "$ref": "#/components/schemas/User" }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID or collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "501": {
            "description": "Collection not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Error updating file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "security": [
        {
          "token_auth": []
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
      "Category": {
        "required": ["name", "category"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24,
            "example": "6208136134ffbeeabc4b1ead"
          },
          "name": {
            "type": "string",
            "example": "PASTRIES"
          },
          "status": {
            "type": "boolean",
            "description": "current status in the store"
          }
        }
      },
...
      "Product": {
        "required": ["name", "category"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24,
            "example": "68376272814ffbeeabc4b1ed"
          },
          "name": {
            "type": "string",
            "example": "PASTRY ROLL"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "status": {
            "type": "boolean",
            "description": "current status in the store"
          }
        }
      },
      "User": {
        "required": ["name", "category"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24,
            "example": "68373827924ffbeeabc4b1ed"
          },
          "name": {
            "type": "string",
            "example": "John Smith"
          },
          "email": {
            "type": "string",
            "example": "test3@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "12345678"
          },
          "img": {
            "type": "string",
            "example": "test.jpg"
          },
          "role": {
            "type": "string",
            "example": "ADMIN_ROLE",
            "enum": ["ADMIN_ROLE", "USER_ROLE"]
          },
          "google": {
            "type": "boolean",
            "description": "has Google authentication"
          },
          "status": {
            "type": "boolean",
            "description": "current status in the store"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "BadRequest"
          },
          "title": {
            "type": "string",
            "example": "Invalid ID"
          },
          "detail": {
            "type": "string",
            "example": "The ID supplied is not a valid MongoID value"
          }
        }
      }
    },
    "securitySchemes": {
      "token_auth": {
        "type": "apiKey",
        "name": "x-token",
        "in": "header"
      }
    }
  }
}